pyproject.toml
[tool.poetry]
name = "converter"
version = "0.1.0"
description = ""
authors = ["Ларькин Михаил <jetpilot314@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "3.14"
python-dotenv = "0.21.1"
black = "^23.7.0"
flake8 = "^6.1.0"
flake8-broken-line = "^1.0.0"
flake8-isort = "^6.0.0"
flake8-plugin-utils = "^1.3.3"
flake8-return = "^1.2.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
